resolve issue - not find categories table, move all the style to style.css to make estimation.html easier to manage
